// types.ts
export interface Snowflake {
  x: number;
  y: number;
  radius: number;
  speed: number;
  wind: number;
}

export interface Star {
  x: number;
  y: number;
  size: number;
  opacity: number;
  twinkleSpeed: number;
}

export interface GiftCardData {
  recipientName: string;
  message: string;
}